requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pymysql==1.1.0
redis==5.0.1
python-redis-lock==4.0.0         
python-jose[cryptography]==3.3.0  
passlib[bcrypt]==1.7.4            
requests==2.31.0
tenacity==8.2.3                   
pydantic==2.4.2
python-multipart==0.0.6
multiprocessing-logging==0.3.4
ratelimit==2.2.1                 
python-dotenv==1.0.0
typing-extensions==4.8.0
DBUtils==3.0.3